refactor auth (many patient methods became universal)
